requirements.txt
beautifulsoup4
requests
pandas
selenium
webdriver_manager
gradio
langchain_community
langchain_openai
langchain_pinecone
langchainhub